Hacks to run on desktop Linux
Not sure what a lot of this was about since it was just random stuff,
but the O3 -> O2 change is necessary, otherwise it wiill crash almost
constantly.

+++ include/liqcell.h
@@ -88,7 +88,7 @@ struct liqcell
 	liqimage  *image;
 	liqfont   *font;
 	int dirty;							// the dirty flag is important, and should really be automatic
-	int tag;
+	intptr_t tag;
 	int dirtyhold;
 	char *caption;
